requirements.txt
# 燧石Agent系统 - Python依赖清单

# 核心依赖
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0

# 向量数据库
chromadb>=0.4.0

# HTTP客户端
requests>=2.31.0
httpx>=0.25.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 配置管理
python-dotenv>=1.0.0
pydantic>=2.0.0

# 日志
loguru>=0.7.0

# 工具
tqdm>=4.65.0
rich>=13.0.0

# Web框架（用于API服务）
fastapi>=0.104.0
uvicorn>=0.24.0

# 异步支持
aiohttp>=3.9.0

# 文档生成
markdown>=3.5.0

# 类型检查
mypy>=1.7.0

# 开发依赖（可选）
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
flake8>=6.1.0
